Move pipeline stats to it's own file
